Tidy up tests and add tests for blinding issuance txs
