+++ pset/pset_test.go
@@ -5,13 +5,15 @@ import (
 	"encoding/hex"
 	"encoding/json"
 	"errors"
-	"github.com/vulpemventures/go-elements/confidential"
+	"fmt"
 	"io/ioutil"
 	"net/http"
 	"os"
 	"strings"
 	"testing"
 	"time"
+
+	"github.com/vulpemventures/go-elements/confidential"
 
 	"github.com/btcsuite/btcd/btcec"
 	"github.com/btcsuite/btcd/txscript"
@@ -77,230 +79,7 @@ func TestRoundTrip(t *testing.T) {
 	}
 }
 
-func TestCreator(t *testing.T) {
-	file, err := ioutil.ReadFile("data/creator.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	var tests []map[string]interface{}
-	err = json.Unmarshal(file, &tests)
-
-	for _, v := range tests {
-		inputs := []*transaction.TxInput{}
-		for _, vIn := range v["inputs"].([]interface{}) {
-			in := vIn.(map[string]interface{})
-			inHash, _ := hex.DecodeString(in["hash"].(string))
-			inIndex := uint32(in["index"].(float64))
-			inHash = bufferutil.ReverseBytes(inHash)
-			inputs = append(inputs, transaction.NewTxInput(inHash, inIndex))
-		}
-
-		outputs := []*transaction.TxOutput{}
-		for _, vOut := range v["outputs"].([]interface{}) {
-			out := vOut.(map[string]interface{})
-			outAsset, _ := hex.DecodeString(out["asset"].(string))
-			outAsset = append([]byte{0x01}, bufferutil.ReverseBytes(outAsset)...)
-			outValue, _ := confidential.SatoshiToElementsValue(uint64(out["value"].(float64)))
-			outScript, _ := hex.DecodeString(out["script"].(string))
-			outputs = append(outputs, transaction.NewTxOutput(outAsset, outValue[:], outScript))
-		}
-
-		p, err := New(inputs, outputs, 2, 0)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		base64Res, err := p.ToBase64()
-		if err != nil {
-			t.Fatal(err)
-		}
-		hexRes, err := p.ToHex()
-		if err != nil {
-			t.Fatal(err)
-		}
-		expectedBase64 := v["expectedBase64"].(string)
-		expectedHex := v["expectedHex"].(string)
-		if base64Res != expectedBase64 {
-			t.Fatalf("Got: %s, expected: %s", base64Res, expectedBase64)
-		}
-		if hexRes != expectedHex {
-			t.Fatalf("Got: %s, expected: %s", hexRes, expectedHex)
-		}
-	}
-}
-
-func TestUpdater(t *testing.T) {
-	file, err := ioutil.ReadFile("data/updater.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	var tests []map[string]interface{}
-	err = json.Unmarshal(file, &tests)
-
-	for _, v := range tests {
-		p, err := NewPsetFromBase64(v["base64"].(string))
-		if err != nil {
-			t.Fatal(err)
-		}
-		updater, err := NewUpdater(p)
-
-		for inIndex, vIn := range v["inputs"].([]interface{}) {
-			in := vIn.(map[string]interface{})
-			if in["nonWitnessUtxo"] != nil {
-				tx, err := transaction.NewTxFromHex(in["nonWitnessUtxo"].(string))
-				if err != nil {
-					t.Fatal(err)
-				}
-				updater.AddInNonWitnessUtxo(tx, inIndex)
-			} else {
-				wu := in["witnessUtxo"].(map[string]interface{})
-				asset, _ := hex.DecodeString(wu["asset"].(string))
-				asset = append([]byte{0x01}, bufferutil.ReverseBytes(asset)...)
-				script, _ := hex.DecodeString(wu["script"].(string))
-				value, _ := confidential.SatoshiToElementsValue(uint64(wu["value"].(float64)))
-				utxo := transaction.NewTxOutput(asset, value[:], script)
-				updater.AddInWitnessUtxo(utxo, inIndex)
-				redeemScript, _ := hex.DecodeString(in["redeemScript"].(string))
-				updater.AddInRedeemScript(redeemScript, inIndex)
-			}
-			updater.AddInSighashType(txscript.SigHashType(int(in["sighashType"].(float64))), inIndex)
-		}
-
-		for outIndex, vOut := range v["outputs"].([]interface{}) {
-			out := vOut.(map[string]interface{})
-			redeemScript, _ := hex.DecodeString(out["redeemScript"].(string))
-			updater.AddOutRedeemScript(redeemScript, outIndex)
-		}
-
-		base64Res, err := updater.Upsbt.ToBase64()
-		if err != nil {
-			t.Fatal(err)
-		}
-		hexRes, err := updater.Upsbt.ToHex()
-		if err != nil {
-			t.Fatal(err)
-		}
-		expectedBase64 := v["expectedBase64"].(string)
-		expectedHex := v["expectedHex"].(string)
-		if base64Res != expectedBase64 {
-			t.Fatalf("Got: %s, expected: %s", base64Res, expectedBase64)
-		}
-		if hexRes != expectedHex {
-			t.Fatalf("Got: %s, expected: %s", hexRes, expectedHex)
-		}
-	}
-}
-
-func TestSigner(t *testing.T) {
-	file, err := ioutil.ReadFile("data/signer.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	var tests []map[string]interface{}
-	err = json.Unmarshal(file, &tests)
-
-	for _, v := range tests {
-		p, err := NewPsetFromBase64(v["base64"].(string))
-		if err != nil {
-			t.Fatal(err)
-		}
-		updater, err := NewUpdater(p)
-
-		for inIndex, vIn := range v["inputs"].([]interface{}) {
-			in := vIn.(map[string]interface{})
-			signature, _ := hex.DecodeString(in["signature"].(string))
-			pubkey, _ := hex.DecodeString(in["pubkey"].(string))
-			updater.Sign(inIndex, signature, pubkey, p.Inputs[inIndex].RedeemScript, p.Inputs[inIndex].WitnessScript)
-		}
-
-		base64Res, err := updater.Upsbt.ToBase64()
-		if err != nil {
-			t.Fatal(err)
-		}
-		hexRes, err := updater.Upsbt.ToHex()
-		if err != nil {
-			t.Fatal(err)
-		}
-		expectedBase64 := v["expectedBase64"].(string)
-		expectedHex := v["expectedHex"].(string)
-		if base64Res != expectedBase64 {
-			t.Fatalf("Got: %s, expected: %s", base64Res, expectedBase64)
-		}
-		if hexRes != expectedHex {
-			t.Fatalf("Got: %s, expected: %s", hexRes, expectedHex)
-		}
-	}
-}
-
-func TestFinalizer(t *testing.T) {
-	file, err := ioutil.ReadFile("data/finalizer.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	var tests []map[string]interface{}
-	err = json.Unmarshal(file, &tests)
-
-	for _, v := range tests {
-		p, err := NewPsetFromBase64(v["base64"].(string))
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		err = FinalizeAll(p)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		base64Res, err := p.ToBase64()
-		if err != nil {
-			t.Fatal(err)
-		}
-		hexRes, err := p.ToHex()
-		if err != nil {
-			t.Fatal(err)
-		}
-		expectedBase64 := v["expectedBase64"].(string)
-		expectedHex := v["expectedHex"].(string)
-		if base64Res != expectedBase64 {
-			t.Fatalf("Got: %s, expected: %s", base64Res, expectedBase64)
-		}
-		if hexRes != expectedHex {
-			t.Fatalf("Got: %s, expected: %s", hexRes, expectedHex)
-		}
-	}
-}
-
-func TestExtractor(t *testing.T) {
-	file, err := ioutil.ReadFile("data/extractor.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	var tests []map[string]interface{}
-	err = json.Unmarshal(file, &tests)
-
-	for _, v := range tests {
-		p, err := NewPsetFromBase64(v["base64"].(string))
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		tx, err := Extract(p)
-		if err != nil {
-			t.Fatal(err)
-		}
-		res, err := tx.ToHex()
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		expectedTxHex := v["expectedTxHex"].(string)
-		if res != expectedTxHex {
-			t.Fatalf("Got: %s, expected: %s", res, expectedTxHex)
-		}
-	}
-}
-
-func TestFromCreateToBroadcast(t *testing.T) {
+func TestBroadcastUnblindedTx(t *testing.T) {
 	/**
 	* This test attempts to broadcast a transaction composed by 1 input and 3
 	* outputs. The input of the transaction will be a native segwit input, thus
@@ -309,9 +88,9 @@ func TestFromCreateToBroadcast(t *testing.T) {
 	* The 3rd output is for the fees, that in Elements side chains are explicits.
 	*
 	* This is intended to test that all methods provided let one to manage a
-	* partial transaction from its creatation to the extraction of the final
-	* tranasction so that it can be correctly broadcasted to the network and
-	* included in the blockchain.
+	* partial unblinded transaction from its creatation to the extraction of
+	* the final tranasction so that it can be correctly broadcasted to the
+	* network and included in the blockchain.
 	 */
 
 	// Generate sender random key pair.
@@ -428,7 +207,7 @@ func TestFromCreateToBroadcast(t *testing.T) {
 	}
 }
 
-func TestCreateAddIssuanceSignAndBroadcast(t *testing.T) {
+func TestBroadcastUnblindedIssuanceTx(t *testing.T) {
 	// Generate sender random key pair.
 	privkey, err := btcec.NewPrivateKey(btcec.S256())
 	if err != nil {
@@ -547,17 +326,367 @@ func TestCreateAddIssuanceSignAndBroadcast(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	res, err := broadcast(txHex)
+	_, err = broadcast(txHex)
+	if err != nil {
+		t.Fatal(err)
+	}
+}
+
+func TestBroadcastBlindedTx(t *testing.T) {
+	/**
+	* This test attempts to broadcast a confidential transaction composed by 1
+	* input and 3 outputs. The input of the transaction will be an unblinded
+	* native segwit input, thus locked by a p2wpkh script, while the outputs
+	* will be a legacy p2sh for the receiver and a segwit p2wpkh for the change.
+	* The 3rd output is for the fees, that in Elements side chains are explicits.
+	*
+	* This is intended to test that all methods provided let one to manage a
+	* partial transaction from its creation to the extraction of the final
+	* tranasction so that it can be correctly broadcasted to the network and
+	* included in the blockchain.
+	 */
+
+	// Generate sender random key pair.
+	privkey, err := btcec.NewPrivateKey(btcec.S256())
+	if err != nil {
+		t.Fatal(err)
+	}
+	pubkey := privkey.PubKey()
+	p2wpkh := payment.FromPublicKey(pubkey, &network.Regtest, nil)
+	address, _ := p2wpkh.WitnessPubKeyHash()
+
+	// Fund sender address.
+	_, err = faucet(address)
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	if len(res) <= 0 || strings.Contains(res, "sendrawtransaction") {
-		t.Fatalf("Expected transaction to be broadcasted, failed for reason: %s", res)
+	// Retrieve sender utxos.
+	utxos, err := unspents(address)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// The transaction will have 1 input and 3 outputs.
+	txInputHash, _ := hex.DecodeString(utxos[0]["txid"].(string))
+	txInputHash = bufferutil.ReverseBytes(txInputHash)
+	txInputIndex := uint32(utxos[0]["vout"].(float64))
+	txInput := transaction.NewTxInput(txInputHash, txInputIndex)
+
+	lbtc, _ := hex.DecodeString("5ac9f65c0efcc4775e0baec4ec03abdde22473cd3cf33c0419ca290e0751b225")
+	lbtc = append([]byte{0x01}, bufferutil.ReverseBytes(lbtc)...)
+	receiverValue, _ := confidential.SatoshiToElementsValue(60000000)
+	receiverScript, _ := hex.DecodeString("76a91439397080b51ef22c59bd7469afacffbeec0da12e88ac")
+	receiverOutput := transaction.NewTxOutput(lbtc, receiverValue[:], receiverScript)
+
+	changeScript := p2wpkh.Script
+	changeValue, _ := confidential.SatoshiToElementsValue(39999500)
+	changeOutput := transaction.NewTxOutput(lbtc, changeValue[:], changeScript)
+
+	feeScript := []byte{}
+	feeValue, _ := confidential.SatoshiToElementsValue(500)
+	feeOutput := transaction.NewTxOutput(lbtc, feeValue[:], feeScript)
+
+	// Create a new pset.
+	inputs := []*transaction.TxInput{txInput}
+	outputs := []*transaction.TxOutput{receiverOutput, changeOutput, feeOutput}
+	p, err := New(inputs, outputs, 2, 0)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Add sighash type and witness utxo to the partial input.
+	updater, err := NewUpdater(p)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	err = updater.AddInSighashType(txscript.SigHashAll, 0)
+	if err != nil {
+		t.Fatal(err)
+	}
+	witValue, _ := confidential.SatoshiToElementsValue(uint64(utxos[0]["value"].(float64)))
+	witnessUtxo := transaction.NewTxOutput(lbtc, witValue[:], p2wpkh.Script)
+	err = updater.AddInWitnessUtxo(witnessUtxo, 0)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	//blind outputs
+	blindingPrivKeys := [][]byte{{}}
+
+	blindingPubKeys := make([][]byte, 0)
+	pk, err := btcec.NewPrivateKey(btcec.S256())
+	if err != nil {
+		t.Fatal(err)
+	}
+	blindingpubkey := pk.PubKey().SerializeCompressed()
+	blindingPubKeys = append(blindingPubKeys, blindingpubkey)
+	pk1, err := btcec.NewPrivateKey(btcec.S256())
+	if err != nil {
+		t.Fatal(err)
+	}
+	blindingpubkey1 := pk1.PubKey().SerializeCompressed()
+	blindingPubKeys = append(blindingPubKeys, blindingpubkey1)
+
+	blinder, err := NewBlinder(
+		p,
+		blindingPrivKeys,
+		blindingPubKeys,
+		nil,
+		nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+	err = blinder.Blind()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// The signing of the input is done by retrieving the proper hash of the serialization
+	// of the transaction (the BIP-0143 segwit version in this case) directly from the pset's
+	// UnsignedTx.
+	// NOTE: to correctly sign an utxo locked by a p2wpkh script, we must use the legacy pubkey script
+	// when serializing the transaction.
+	legacyScript := append(append([]byte{0x76, 0xa9, 0x14}, p2wpkh.Hash...), []byte{0x88, 0xac}...)
+	witHash := updater.Upsbt.UnsignedTx.HashForWitnessV0(0, legacyScript, witValue[:], txscript.SigHashAll)
+	sig, err := privkey.Sign(witHash[:])
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	sigWithHashType := append(sig.Serialize(), byte(txscript.SigHashAll))
+
+	// Update the pset adding the input signature script and the pubkey.
+	_, err = updater.Sign(0, sigWithHashType, pubkey.SerializeCompressed(), nil, nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Finalize the partial transaction.
+	p = updater.Upsbt
+	err = FinalizeAll(p)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Extract the final signed transaction from the Pset wrapper.
+	finalTx, err := Extract(p)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Serialize the transaction and try to broadcast.
+	txHex, err := finalTx.ToHex()
+	if err != nil {
+		t.Fatal(err)
+	}
+	_, err = broadcast(txHex)
+	if err != nil {
+		t.Fatal(err)
 	}
 }
 
-func TestCreateBlindIssuanceAndOutputSignAndBroadcast(t *testing.T) {
+func TestBroadcastBlindedTxWithBlindedInput(t *testing.T) {
+	/**
+	* This test attempts to broadcast a confidential transaction composed by 1
+	* confidential input and 3 outputs.
+	* The input of the transaction will be a blinded native segwit input, thus
+	* locked by a p2wpkh script, while the outputs will be a legacy p2sh for the
+	* receiver and a segwit p2wpkh for the change.
+	* The 3rd output is for the fees, that in Elements side chains are explicits.
+	*
+	* This is intended to test that all methods provided let one to manage a
+	* partial transaction from its creatation to the extraction of the final
+	* tranasction so that it can be correctly broadcasted to the network and
+	* included in the blockchain.
+	 */
+
+	// Generate sender random key pair.
+	privkey, err := btcec.NewPrivateKey(btcec.S256())
+	if err != nil {
+		t.Fatal(err)
+	}
+	pubkey := privkey.PubKey()
+
+	blindingPrivateKey, err := btcec.NewPrivateKey(btcec.S256())
+	if err != nil {
+		t.Fatal(err)
+	}
+	blindingPublicKey := blindingPrivateKey.PubKey()
+
+	p2wpkh := payment.FromPublicKey(pubkey, &network.Regtest, blindingPublicKey)
+
+	confidentialAddress, _ := p2wpkh.ConfidentialWitnessPubKeyHash()
+
+	// Fund sender address.
+	_, err = faucet(confidentialAddress)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Retrieve sender utxos.
+	utxos, err := unspents(confidentialAddress)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// The transaction will have 1 input and 3 outputs.
+	txInputHash, _ := hex.DecodeString(utxos[0]["txid"].(string))
+	txInputHash = bufferutil.ReverseBytes(txInputHash)
+	txInputIndex := uint32(utxos[0]["vout"].(float64))
+	txInput := transaction.NewTxInput(txInputHash, txInputIndex)
+
+	lbtc, _ := hex.DecodeString("5ac9f65c0efcc4775e0baec4ec03abdde22473cd3cf33c0419ca290e0751b225")
+	lbtc = append([]byte{0x01}, bufferutil.ReverseBytes(lbtc)...)
+	receiverValue, _ := confidential.SatoshiToElementsValue(60000000)
+	receiverScript, _ := hex.DecodeString("76a91439397080b51ef22c59bd7469afacffbeec0da12e88ac")
+	receiverOutput := transaction.NewTxOutput(lbtc, receiverValue[:], receiverScript)
+
+	changeScript := p2wpkh.Script
+	changeValue, _ := confidential.SatoshiToElementsValue(39999500)
+	changeOutput := transaction.NewTxOutput(lbtc, changeValue[:], changeScript)
+
+	feeScript := []byte{}
+	feeValue, _ := confidential.SatoshiToElementsValue(500)
+	feeOutput := transaction.NewTxOutput(lbtc, feeValue[:], feeScript)
+
+	// Create a new pset.
+	inputs := []*transaction.TxInput{txInput}
+	outputs := []*transaction.TxOutput{receiverOutput, changeOutput, feeOutput}
+	p, err := New(inputs, outputs, 2, 0)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Add sighash type and witness utxo to the partial input.
+	updater, err := NewUpdater(p)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	err = updater.AddInSighashType(txscript.SigHashAll, 0)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	tx, err := fetchTx(utxos[0]["txid"].(string))
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	trx, err := transaction.NewTxFromHex(string(tx))
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	valueCommitment, err := hex.DecodeString(utxos[0]["valuecommitment"].(string))
+	if err != nil {
+		t.Fatal(err)
+	}
+	assetCommitment, err := hex.DecodeString(utxos[0]["assetcommitment"].(string))
+	if err != nil {
+		t.Fatal(err)
+	}
+	witnessUtxo := &transaction.TxOutput{
+		Asset:           assetCommitment,
+		Value:           valueCommitment,
+		Script:          p2wpkh.Script,
+		Nonce:           trx.Outputs[txInputIndex].Nonce,
+		RangeProof:      trx.Outputs[txInputIndex].RangeProof,
+		SurjectionProof: trx.Outputs[txInputIndex].SurjectionProof,
+	}
+	err = updater.AddInWitnessUtxo(witnessUtxo, 0)
+	if err != nil {
+		t.Fatal(err)
+	}
+	//blind outputs
+	blindingPrivKeys := [][]byte{blindingPrivateKey.Serialize()}
+
+	blindingPubKeys := make([][]byte, 0)
+	pk, err := btcec.NewPrivateKey(btcec.S256())
+	if err != nil {
+		t.Fatal(err)
+	}
+	blindingpubkey := pk.PubKey().SerializeCompressed()
+	blindingPubKeys = append(blindingPubKeys, blindingpubkey)
+	pk1, err := btcec.NewPrivateKey(btcec.S256())
+	if err != nil {
+		t.Fatal(err)
+	}
+	blindingpubkey1 := pk1.PubKey().SerializeCompressed()
+	blindingPubKeys = append(blindingPubKeys, blindingpubkey1)
+
+	blinder, err := NewBlinder(
+		p,
+		blindingPrivKeys,
+		blindingPubKeys,
+		nil,
+		nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+	err = blinder.Blind()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// The signing of the input is done by retrieving the proper hash of the
+	//serialization of the transaction (the BIP-0143 segwit version in this
+	//case) directly from the pset's UnsignedTx.
+	//NOTE: to correctly sign an utxo locked by a p2wpkh script,
+	//we must use the legacy pubkey script when serializing the transaction.
+	legacyScript := append(
+		append(
+			[]byte{0x76, 0xa9, 0x14},
+			p2wpkh.Hash...,
+		),
+		[]byte{0x88, 0xac}...,
+	)
+	witHash := updater.Upsbt.UnsignedTx.HashForWitnessV0(
+		0,
+		legacyScript,
+		witnessUtxo.Value,
+		txscript.SigHashAll,
+	)
+	sig, err := privkey.Sign(witHash[:])
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	sigWithHashType := append(sig.Serialize(), byte(txscript.SigHashAll))
+
+	// Update the pset adding the input signature script and the pubkey.
+	_, err = updater.Sign(0, sigWithHashType, pubkey.SerializeCompressed(), nil, nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Finalize the partial transaction.
+	p = updater.Upsbt
+	err = FinalizeAll(p)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Extract the final signed transaction from the Pset wrapper.
+	finalTx, err := Extract(p)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Serialize the transaction and try to broadcast.
+	txHex, err := finalTx.ToHex()
+	if err != nil {
+		t.Fatal(err)
+	}
+	_, err = broadcast(txHex)
+	if err != nil {
+		t.Fatal(err)
+	}
+}
+
+func TestBroadcastIssuanceTxWithBlindedOutput(t *testing.T) {
 	// Generate sender random key pair.
 	privkey, err := btcec.NewPrivateKey(btcec.S256())
 	if err != nil {
@@ -585,21 +714,14 @@ func TestCreateBlindIssuanceAndOutputSignAndBroadcast(t *testing.T) {
 	txInputIndex := uint32(utxos[0]["vout"].(float64))
 	txInput := transaction.NewTxInput(txInputHash, txInputIndex)
 
-	lbtc, _ := hex.DecodeString("5ac9f65c0efcc4775e0baec4ec03abdde22473cd3" +
-		"cf33c0419ca290e0751b225")
+	lbtc, _ := hex.DecodeString(
+		"5ac9f65c0efcc4775e0baec4ec03abdde22473cd3cf33c0419ca290e0751b225",
+	)
 	lbtc = append([]byte{0x01}, bufferutil.ReverseBytes(lbtc)...)
-
-	changeScript := p2wpkh.Script
-	changeValue, _ := confidential.SatoshiToElementsValue(99999500)
-	changeOutput := transaction.NewTxOutput(lbtc, changeValue[:], changeScript)
-
-	feeScript := []byte{}
-	feeValue, _ := confidential.SatoshiToElementsValue(500)
-	feeOutput := transaction.NewTxOutput(lbtc, feeValue[:], feeScript)
 
 	// Create a new pset.
 	inputs := []*transaction.TxInput{txInput}
-	outputs := []*transaction.TxOutput{changeOutput, feeOutput}
+	outputs := []*transaction.TxOutput{}
 	p, err := New(inputs, outputs, 2, 0)
 	if err != nil {
 		t.Fatal(err)
@@ -644,31 +766,39 @@ func TestCreateBlindIssuanceAndOutputSignAndBroadcast(t *testing.T) {
 		t.Fatal(err)
 	}
 
+	// Add change and fees
+	changeScript := p2wpkh.Script
+	changeValue, _ := confidential.SatoshiToElementsValue(99996000)
+	changeOutput := transaction.NewTxOutput(lbtc, changeValue[:], changeScript)
+	updater.AddOutput(changeOutput)
+
+	feeScript := []byte{}
+	feeValue, _ := confidential.SatoshiToElementsValue(4000)
+	feeOutput := transaction.NewTxOutput(lbtc, feeValue[:], feeScript)
+	updater.AddOutput(feeOutput)
+
 	//blind outputs
 	blindingPrivKeys := [][]byte{{}}
-
 	blindingPubKeys := make([][]byte, 0)
 
 	pk, err := btcec.NewPrivateKey(btcec.S256())
 	if err != nil {
 		t.Fatal(err)
 	}
+	// blinding pubkey for asset output
 	blindingpubkey := pk.PubKey().SerializeCompressed()
 	blindingPubKeys = append(blindingPubKeys, blindingpubkey)
 
+	// blinding pubkey for token output
+	blindingPubKeys = append(blindingPubKeys, blindingpubkey)
+
+	// blinding pubkey for change output
 	pk1, err := btcec.NewPrivateKey(btcec.S256())
 	if err != nil {
 		t.Fatal(err)
 	}
 	blindingpubkey1 := pk1.PubKey().SerializeCompressed()
 	blindingPubKeys = append(blindingPubKeys, blindingpubkey1)
-
-	pk2, err := btcec.NewPrivateKey(btcec.S256())
-	if err != nil {
-		t.Fatal(err)
-	}
-	blindingpubkey2 := pk2.PubKey().SerializeCompressed()
-	blindingPubKeys = append(blindingPubKeys, blindingpubkey2)
 
 	issuanceBlindingPrivateKeys := make([]IssuanceBlindingPrivateKeys, 0)
 	issuanceBlindingPrivateKeys = append(
@@ -684,11 +814,12 @@ func TestCreateBlindIssuanceAndOutputSignAndBroadcast(t *testing.T) {
 		blindingPrivKeys,
 		blindingPubKeys,
 		nil,
-		issuanceBlindingPrivateKeys)
+		nil, // issuanceBlindingPrivateKeys,
+	)
 	if err != nil {
 		t.Fatal(err)
 	}
-	err = blinder.BlindTransaction()
+	err = blinder.Blind()
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -726,13 +857,179 @@ func TestCreateBlindIssuanceAndOutputSignAndBroadcast(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	res, err := broadcast(txHex)
+
+	_, err = broadcast(txHex)
+	if err != nil {
+		t.Fatal(err)
+	}
+}
+
+func TestBroadcastBlindedIssuanceTx(t *testing.T) {
+	// Generate sender random key pair.
+	privkey, err := btcec.NewPrivateKey(btcec.S256())
+	if err != nil {
+		t.Fatal(err)
+	}
+	pubkey := privkey.PubKey()
+	p2wpkh := payment.FromPublicKey(pubkey, &network.Regtest, nil)
+	address, _ := p2wpkh.WitnessPubKeyHash()
+
+	// Fund sender address.
+	_, err = faucet(address)
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	if len(res) <= 0 || strings.Contains(res, "sendrawtransaction") {
-		t.Fatalf("Expected transaction to be broadcasted, failed for reason: %s", res)
+	// Retrieve sender utxos.
+	utxos, err := unspents(address)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// The transaction will have 1 input and 2 outputs.
+	txInputHash, _ := hex.DecodeString(utxos[0]["txid"].(string))
+	txInputHash = bufferutil.ReverseBytes(txInputHash)
+	txInputIndex := uint32(utxos[0]["vout"].(float64))
+	txInput := transaction.NewTxInput(txInputHash, txInputIndex)
+
+	lbtc, _ := hex.DecodeString(
+		"5ac9f65c0efcc4775e0baec4ec03abdde22473cd3cf33c0419ca290e0751b225",
+	)
+	lbtc = append([]byte{0x01}, bufferutil.ReverseBytes(lbtc)...)
+
+	// Create a new pset.
+	inputs := []*transaction.TxInput{txInput}
+	outputs := []*transaction.TxOutput{}
+	p, err := New(inputs, outputs, 2, 0)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	updater, err := NewUpdater(p)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	arg := AddIssuanceArg{
+		Precision:    0,
+		AssetAmount:  2000,
+		TokenAmount:  0.5,
+		AssetAddress: address,
+		TokenAddress: address,
+		TokenFlag:    1,
+		Net:          network.Regtest,
+	}
+	err = updater.AddIssuance(arg)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	err = updater.AddInSighashType(txscript.SigHashAll, 0)
+	if err != nil {
+		t.Fatal(err)
+	}
+	witValue, _ := confidential.SatoshiToElementsValue(uint64(utxos[0]["value"].(float64)))
+	witnessUtxo := transaction.NewTxOutput(lbtc, witValue[:], p2wpkh.Script)
+	err = updater.AddInWitnessUtxo(witnessUtxo, 0)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Add change and fees
+	changeScript := p2wpkh.Script
+	changeValue, _ := confidential.SatoshiToElementsValue(99996000)
+	changeOutput := transaction.NewTxOutput(lbtc, changeValue[:], changeScript)
+	updater.AddOutput(changeOutput)
+
+	feeScript := []byte{}
+	feeValue, _ := confidential.SatoshiToElementsValue(4000)
+	feeOutput := transaction.NewTxOutput(lbtc, feeValue[:], feeScript)
+	updater.AddOutput(feeOutput)
+
+	//blind outputs
+	blindingPrivKeys := [][]byte{{}}
+	blindingPubKeys := make([][]byte, 0)
+
+	pk, err := btcec.NewPrivateKey(btcec.S256())
+	if err != nil {
+		t.Fatal(err)
+	}
+	// blinding pubkey for asset output
+	blindingpubkey := pk.PubKey().SerializeCompressed()
+	blindingPubKeys = append(blindingPubKeys, blindingpubkey)
+
+	// blinding pubkey for token output
+	blindingPubKeys = append(blindingPubKeys, blindingpubkey)
+
+	// blinding pubkey for change output
+	pk1, err := btcec.NewPrivateKey(btcec.S256())
+	if err != nil {
+		t.Fatal(err)
+	}
+	blindingpubkey1 := pk1.PubKey().SerializeCompressed()
+	blindingPubKeys = append(blindingPubKeys, blindingpubkey1)
+
+	issuanceBlindingPrivateKeys := make([]IssuanceBlindingPrivateKeys, 0)
+	issuanceBlindingPrivateKeys = append(
+		issuanceBlindingPrivateKeys,
+		IssuanceBlindingPrivateKeys{
+			AssetKey: pk.Serialize(),
+			TokenKey: pk1.Serialize(),
+		},
+	)
+
+	blinder, err := NewBlinder(
+		p,
+		blindingPrivKeys,
+		blindingPubKeys,
+		nil,
+		issuanceBlindingPrivateKeys,
+	)
+	if err != nil {
+		t.Fatal(err)
+	}
+	err = blinder.Blind()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	legacyScript := append(append([]byte{0x76, 0xa9, 0x14}, p2wpkh.Hash...), []byte{0x88, 0xac}...)
+	witHash := updater.Upsbt.UnsignedTx.HashForWitnessV0(0, legacyScript, witValue[:], txscript.SigHashAll)
+	sig, err := privkey.Sign(witHash[:])
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	sigWithHashType := append(sig.Serialize(), byte(txscript.SigHashAll))
+
+	// Update the pset adding the input signature script and the pubkey.
+	_, err = updater.Sign(0, sigWithHashType, pubkey.SerializeCompressed(), nil, nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Finalize the partial transaction.
+	p = updater.Upsbt
+	err = FinalizeAll(p)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Extract the final signed transaction from the Pset wrapper.
+	finalTx, err := Extract(p)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Serialize the transaction and try to broadcast.
+	txHex, err := finalTx.ToHex()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	_, err = broadcast(txHex)
+	if err != nil {
+		t.Fatal(err)
 	}
 }
 
@@ -817,5 +1114,28 @@ func broadcast(txHex string) (string, error) {
 		return "", err
 	}
 
-	return string(txid), nil
+	res := string(txid)
+	if len(res) <= 0 || strings.Contains(res, "sendrawtransaction") {
+		return "", fmt.Errorf("Expected transaction to be broadcasted, failed for reason: %s", res)
+	}
+
+	return res, nil
+}
+
+func fetchTx(txId string) ([]byte, error) {
+	baseUrl, ok := os.LookupEnv("API_URL")
+	if !ok {
+		return nil, errors.New("API_URL environment variable is not set")
+	}
+	url := baseUrl + "/tx/" + txId + "/hex"
+	resp, err := http.Get(url)
+	if err != nil {
+		return nil, err
+	}
+	data, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return nil, err
+	}
+
+	return data, nil
 }
